+++ pkgs/ffigen/lib/src/config_provider/yaml_config.dart
@@ -219,10 +219,6 @@ final class YamlConfig {
   String? get preamble => _preamble;
   String? _preamble;
 
-  /// If `Dart_Handle` should be mapped with Handle/Object.
-  bool get useDartHandle => _useDartHandle;
-  late bool _useDartHandle;
-
   /// Where to silence warning for enum integer type mimicking.
   bool get silenceEnumWarning => _silenceEnumWarning;
   late bool _silenceEnumWarning;
@@ -878,12 +874,6 @@ final class YamlConfig {
           ),
         ),
         HeterogeneousMapEntry(
-          key: strings.useDartHandle,
-          valueConfigSpec: BoolConfigSpec(),
-          defaultValue: (node) => true,
-          resultOrDefault: (node) => _useDartHandle = node.value as bool,
-        ),
-        HeterogeneousMapEntry(
           key: strings.ffiNative,
           valueConfigSpec: OneOfConfigSpec(
             childConfigSpecs: [
@@ -1320,7 +1310,6 @@ final class YamlConfig {
         ? ObjectiveC(
             interfaces: Interfaces(
               include: objcInterfaces.shouldInclude,
-              includeSymbolAddress: objcInterfaces.shouldIncludeSymbolAddress,
               includeMember: objcInterfaces.shouldIncludeMember,
               rename: objcInterfaces.rename,
               renameMember: objcInterfaces.renameMember,
@@ -1329,7 +1318,6 @@ final class YamlConfig {
             ),
             protocols: Protocols(
               include: objcProtocols.shouldInclude,
-              includeSymbolAddress: objcProtocols.shouldIncludeSymbolAddress,
               includeMember: objcProtocols.shouldIncludeMember,
               rename: objcProtocols.rename,
               renameMember: objcProtocols.renameMember,
@@ -1338,7 +1326,6 @@ final class YamlConfig {
             ),
             categories: Categories(
               include: objcCategories.shouldInclude,
-              includeSymbolAddress: objcCategories.shouldIncludeSymbolAddress,
               includeMember: objcCategories.shouldIncludeMember,
               rename: objcCategories.rename,
               renameMember: objcCategories.renameMember,
@@ -1355,8 +1342,6 @@ final class YamlConfig {
     importedTypesByUsr: usrTypeMappings,
     // ignore: deprecated_member_use_from_same_package
     integers: Integers(imported: nativeTypeMappings.values.toList()),
-    // ignore: deprecated_member_use_from_same_package
-    useDartHandle: useDartHandle,
     // ignore: deprecated_member_use_from_same_package
     libclangDylib: libclangDylib,
   );