+++ pkgs/swiftgen/lib/src/generator.dart
@@ -98,7 +98,6 @@ extension SwiftGenGenerator on SwiftGenerator {
         interfaces: fg.Interfaces(
           include: interfaces.include,
           includeMember: interfaces.includeMember,
-          includeSymbolAddress: interfaces.includeSymbolAddress,
           rename: interfaces.rename,
           renameMember: interfaces.renameMember,
           includeTransitive: interfaces.includeTransitive,
@@ -109,7 +108,6 @@ extension SwiftGenGenerator on SwiftGenerator {
         protocols: fg.Protocols(
           include: protocols.include,
           includeMember: protocols.includeMember,
-          includeSymbolAddress: protocols.includeSymbolAddress,
           rename: protocols.rename,
           renameMember: protocols.renameMember,
           includeTransitive: protocols.includeTransitive,